remove code dead
