+++ package.json
@@ -6,14 +6,10 @@
     "@material-ui/core": "^4.9.14",
     "@material-ui/icons": "^4.9.1",
     "@material-ui/lab": "^4.0.0-alpha.53",
-    "@material-ui/pickers": "^3.2.10",
     "@material-ui/styles": "^4.9.14",
     "@testing-library/react": "^9.3.2",
-    "@testing-library/user-event": "^7.1.2",
     "formik": "^2.1.4",
     "camelcase-keys": "^6.2.2",
-    "highcharts": "^8.1.0",
-    "highcharts-react-official": "^3.0.0",
     "polished": "^3.6.3",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
@@ -50,7 +46,6 @@
     "@types/node": "^12.0.0",
     "@types/react": "^16.9.0",
     "@types/react-dom": "^16.9.0",
-    "@types/react-highcharts": "^16.0.2",
     "@types/react-router-dom": "^5.1.5"
   }
 }
