+++ src/components/Header.tsx
@@ -1,40 +1,39 @@
-import { AppBar, Toolbar, Typography } from '@material-ui/core';
-import { createStyles, makeStyles } from '@material-ui/styles';
-import React from 'react';
-import { NavLink } from 'react-router-dom';
-import { userLogOut } from '../services/Auth';
+import { AppBar, Button, Toolbar, Typography } from "@material-ui/core";
+import { createStyles, makeStyles } from "@material-ui/styles";
+import React, { useCallback } from "react";
+import { useHistory } from "react-router-dom";
 
 export const Header = () => {
-    const classes = useStyles();
+  const classes = useStyles();
+  const { push } = useHistory();
 
-    return (
-        <AppBar position="static">
-            <Toolbar>
-                <Typography variant="h6" className={classes.title}>
-                    Nemesis
-                </Typography>
-                <NavLink to="/login" className={classes.logout}onClick={() => userLogOut()} >
-                    <span onClick={() => userLogOut()}>Logout</span>
-                </NavLink>
-            </Toolbar>
-        </AppBar>
-    )
-}
+  const handleOnClick = useCallback(() => {
+    push("/addUsers");
+  }, [push]);
+
+  return (
+    <AppBar position="static" color="default">
+      <Toolbar>
+        <Typography variant="h6" className={classes.title}>
+          Mexican Test
+        </Typography>
+        <Button
+          variant="contained"
+          size="small"
+          color="primary"
+          onClick={handleOnClick}
+        >
+          Create User
+        </Button>
+      </Toolbar>
+    </AppBar>
+  );
+};
 
 const useStyles = makeStyles(() =>
-    createStyles({
-        link: {
-            textDecoration: 'none',
-            color: 'inherit'
-        },
-        title: {
-            flexGrow: 1,
-        },
-        logout:{
-            textDecoration: 'none',
-            color: 'inherit',
-            paddingLeft:20,
-        }
-    }),
+  createStyles({
+    title: {
+      flexGrow: 1,
+    },
+  })
 );
-
