--- src/helpers/requestHostInterceptor.ts
@@ -1,8 +1,0 @@
-import { RequestInterceptor } from 'react-fetching-library';
-
-export const requestHostInterceptor: (host: string) => RequestInterceptor = host => () => async action => {
-        return {
-                ...action,
-                endpoint: `${host}${action.endpoint}`,
-        };
-};
