+++ src/routes/route.tsx
@@ -1,14 +1,12 @@
-import React from 'react'
-import { Switch, Route, Redirect } from "react-router-dom";
-import Login from '../pages/Login';
-import Users from '../pages/Users';
+import React from 'react';
+import { Route, Switch } from "react-router-dom";
 import { AddUserForm } from '../AddUserForm';
 import Dashboard from '../pages/Dashboard';
+import Users from '../pages/Users';
 
 export const Router = () => {
     return (
         <Switch>
-            {/* <Route exact path="/" component={Login} /> */}
             <Route exact path="/" component={Dashboard}/>
             <Route path="/users" component={Users} />
             <Route path="/addUsers" component={AddUserForm} />
