--- src/types.ts
@@ -1,6 +1,0 @@
-export type User = {
-  id: number;
-  name: string;
-  keyword: string;
-  roll: string;
-};
