+++ src/utils.ts
@@ -1,3 +1,5 @@
+import { Action } from 'react-fetching-library';
+
 export const apiPath = process.env.REACT_APP_API || "https://reqres.in/api";
 
 export const getEndPointPath = (path: string, search?: string) =>
